requirements.txt
fastapi
uvicorn
streamlit
nsetools
feedparser
sentence-transformers
faiss-cpu
transformers
gtts
playsound
openai-whisper
pydub
